requirements.txt
Flask==3.1.3
implicit==0.7.2
numpy==2.4.3
pandas==3.0.1
pyarrow==23.0.1
PyMySQL==1.1.2
python-dotenv==1.2.2
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
SQLAlchemy==2.0.48